requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.6
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-xml==2.0.0
djoser==2.2.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
urllib3==1.26.15
